Minor update
++ Changed ++
- createDisk was changed from uint32_t parameter to int64_t
- run script to ensure a clean terminal output during compilation

++ Added ++
- File size check during creation
- File existence check during loading
- Single loadable files

+++ code/include/disk.h
@@ -81,7 +81,7 @@ public methods:
      *          virtual disk is determined by @see{#size}
      * @param size Size of the disk in bytes
      */
-    void createDisk(uint32_t size_m);
+    void createDisk(int64_t size_m);
 
     void deleteDisk();
